Separate details from public API
