+++ test/extra/persist/test_conversion.cpp
@@ -1,8 +1,8 @@
 #include <catch2/catch_test_macros.hpp>
 
-#include <immer/extra/persist/json/json_with_pool_auto.hpp>
-
 #include <immer/extra/persist/detail/type_traverse.hpp>
+#include <immer/extra/persist/json/json_with_pool.hpp>
+#include <immer/extra/persist/transform.hpp>
 
 #include "utils.hpp"
