+++ test/extra/persist/test_for_docs.cpp
@@ -1,6 +1,6 @@
 #include <catch2/catch_test_macros.hpp>
 
-#include <immer/extra/persist/json/json_with_pool_auto.hpp>
+#include <immer/extra/persist/json/json_with_pool.hpp>
 
 #include "utils.hpp"
 #include <nlohmann/json.hpp>
